地震速報をの処理の部分を、クラスを継承してオーバーライドする形式に変更。

+++ jparser.py
@@ -140,7 +140,7 @@ class EqHypocenter(EqBase):
 
     @property
     def magnitude(self):
-        return int(self.magnitude_raw)
+        return float(self.magnitude_raw)
 
     @property
     def magnitude_text(self):
